fix(avalonia): address PR #1481 review — explicit version-6 branch, ulong slot arithmetic, test casts
Review-thread fixes for #1420:
- WorldMapEventResolverCore.GetEventByMapID: handle version==6 explicitly and
  return NOT_FOUND for any other version (ROMFE0/version 0 or future), instead
  of falling through to the FE6 PLIST path and reading unrelated pointers.
- ResolveIndexedTable: compute the table slot in ulong so a large/corrupted
  wmapid cannot overflow/wrap a uint past the bounds check.
- Tests: cast maps.Count to uint in the for-loops (explicit, avoids any
  int/uint comparison ambiguity).

New tests: UnknownVersion_NotFound (ROMFE0) + FE8_HugeMapId_NoOverflowWrap_NotFound.
All 16 Core + 6 Avalonia tests pass.

+++ FEBuilderGBA.Avalonia.Tests/ToolExportEAEventWorldMapTests.cs
@@ -65,7 +65,7 @@ namespace FEBuilderGBA.Avalonia.Tests
                 uint firstAddr = U.NOT_FOUND;
                 uint firstMap = 0;
                 var maps = MapSettingCore.MakeMapIDList(rom);
-                for (uint mapId = 0; mapId < maps.Count; mapId++)
+                for (uint mapId = 0; mapId < (uint)maps.Count; mapId++)
                 {
                     uint mapAddr = MapSettingCore.GetMapAddr(rom, mapId);
                     if (!U.isSafetyOffset(mapAddr, rom)) continue;
